+++ HISTORY.rst
@@ -7,6 +7,9 @@ History
 
 * Better handling for a large variety of error conditions
 * Proper fallback mechanism implemented
+* Various bug fixes:
+  * Avoid output ``.elm`` files with no exports
+  * Bugs with HTML attributes and non-string message args
 
 
 0.1.0 (2018-07-27)