+++ HISTORY.rst
@@ -9,8 +9,9 @@ History
 * Proper fallback mechanism implemented
 * Added ``--watch`` option.
 * Various bug fixes:
-  * Avoid output ``.elm`` files with no exports
+  * Avoid outputting ``.elm`` files with no exports
   * Bugs with HTML attributes and non-string message args
+  * Crasher with multi-line messages
 
 
 0.1.0 (2018-07-27)